+++ ngmix/priors/constants.py
@@ -1,0 +1,4 @@
+import numpy
+
+LOWVAL = -numpy.inf
+BIGVAL = 9999.0e47
